Try to stabilize more tests by waiting for publisher
